Builds with foobar2000 SDK 2023-09-06

+++ Preferences.cpp
@@ -140,10 +140,10 @@ private:
         _DarkModeHooks.AddDialogWithControls(*this);
 
         _SamplesPath = CfgSamplesPath;
-        _PlaybackMode = CfgPlaybackMode;
-        _LoopCount = CfgLoopCount;
-        _FadeOutDuration = CfgFadeOutDuration;
-        _SynthesisRate = CfgSynthesisRate;
+        _PlaybackMode = (uint32_t) CfgPlaybackMode;
+        _LoopCount = (uint32_t) CfgLoopCount;
+        _FadeOutDuration = (uint32_t) CfgFadeOutDuration;
+        _SynthesisRate = (uint32_t) CfgSynthesisRate;
         _UsePPS = CfgUsePPS;
         _UseSSG = CfgUseSSG;
